modify RocketmqTopicCore, using SDKManager
