+++ eventmesh-dashboard-console/src/main/java/org/apache/eventmesh/dashboard/console/controller/TopicController.java
@@ -18,7 +18,7 @@
 package org.apache.eventmesh.dashboard.console.controller;
 
 import org.apache.eventmesh.dashboard.common.dto.Result;
-import org.apache.eventmesh.dashboard.common.model.TopicProperties;
+import org.apache.eventmesh.dashboard.service.dto.TopicProperties;
 import org.apache.eventmesh.dashboard.console.dto.CreateTopicRequest;
 import org.apache.eventmesh.dashboard.console.dto.DeleteTopicRequest;
 import org.apache.eventmesh.dashboard.service.store.TopicCore;
@@ -59,14 +59,14 @@ public class TopicController {
     }
 
     @CrossOrigin
-    @GetMapping
+    @GetMapping("/getList")
     public Result<List<TopicProperties>> getList() {
-        List<TopicProperties> topicList = topicCore.getTopic();
+        List<TopicProperties> topicList = topicCore.getTopics();
         return Result.success(topicList);
     }
 
     @CrossOrigin
-    @PostMapping
+    @PostMapping("/create")
     public Result<Object> create(@RequestBody CreateTopicRequest createTopicRequest) {
         String topicName = createTopicRequest.getName();
         topicCore.createTopic(topicName);
@@ -74,7 +74,7 @@ public class TopicController {
     }
 
     @CrossOrigin
-    @DeleteMapping
+    @DeleteMapping("delete")
     public Result<Object> delete(@RequestBody DeleteTopicRequest deleteTopicRequest) {
         String topicName = deleteTopicRequest.getName();
         topicCore.deleteTopic(topicName);
