+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/function/SDK/AbstractSDKOperation.java
@@ -15,21 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.dashboard.service.store;
+package org.apache.eventmesh.dashboard.core.function.SDK;
 
-import org.apache.eventmesh.dashboard.service.dto.TopicProperties;
+public abstract class AbstractSDKOperation<T> implements SDKOperation<T> {
 
-import java.util.List;
-
-/**
- * Manage topics of eventmesh-storage-plugin (EventMesh Store).
- */
-
-public interface TopicCore {
-
-    List<TopicProperties> getTopics();
-
-    Boolean createTopic(String topicName);
-
-    Boolean deleteTopic(String topicName);
+    protected T castClient(Object client) {
+        try {
+            return (T) client;
+        } catch (ClassCastException e) {
+            throw new IllegalArgumentException("Client is not of the expected type", e);
+        }
+    }
 }
