+++ eventmesh-dashboard-service/src/main/java/org/apache/eventmesh/dashboard/service/dto/TopicProperties.java
@@ -15,21 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.dashboard.service.store;
+package org.apache.eventmesh.dashboard.service.dto;
 
-import org.apache.eventmesh.dashboard.service.dto.TopicProperties;
+import org.apache.rocketmq.common.TopicFilterType;
 
-import java.util.List;
+import lombok.Data;
 
 /**
- * Manage topics of eventmesh-storage-plugin (EventMesh Store).
+ * One record displayed in 'Topic' page.
  */
 
-public interface TopicCore {
+@Data
+public class TopicProperties {
 
-    List<TopicProperties> getTopics();
+    private String topicName;
 
-    Boolean createTopic(String topicName);
+    private int offset;
 
-    Boolean deleteTopic(String topicName);
+    private int readQueueNums;
+
+    private int writeQueueNums;
+
 }
